[STCK-002] Refactoring: quote creation logic moved to a separate package (service layer).
