+++ src/main/kotlin/io/stocks/inc/management/service/quote/creation/StockCreationRequestService.kt
@@ -1,4 +1,4 @@
-package io.stocks.inc.management.service
+package io.stocks.inc.management.service.quote.creation
 
 import io.stocks.inc.management.extensions.toDto
 import io.stocks.inc.management.extensions.toModelWithPeriodIdProvider
