+++ src/main/kotlin/io/stocks/inc/management/service/quote/creation/StockQuoteSavingHandler.kt
@@ -1,4 +1,4 @@
-package io.stocks.inc.management.service
+package io.stocks.inc.management.service.quote.creation
 
 import io.stocks.inc.management.model.StockQuoteEntry
 import io.stocks.inc.management.model.StockQuoteRequest
